requirements.txt
Flask
PyPDF2
sentence-transformers
faiss-cpu
openai
google-genai
python-dotenv
